adding a disc texture to have a nice icon
as it looks like loading .png files as textures does not really
work as expected, it has been preconverted to an RGBA32 texture
with an external tool.

format of the RGBA32 texture is just 256x256 pixels with a
8 bits each color (so 4 bytes), so writing a tool for that is
simple, if we ever want to update it.

it might make sense to add size information as a header to the
texture later, but that should be fine for now.